Add prisma store for channels and refactoring
