+++ slackapi-app/actions/index.js
@@ -1,0 +1,11 @@
+const { onClickSetChannelSignButton } = require('./conbot/onClickSetChannelSignButton')
+const { onSelectChannel } = require('./conbot/onSelectChannel')
+
+const initActions = ({channels, setReactionAddedMode}) => {
+    onClickSetChannelSignButton(channels)
+    onSelectChannel(setReactionAddedMode)
+}
+
+module.exports = {
+    initActions
+}
