+++ slackapi-app/cach.js
@@ -1,0 +1,6 @@
+const NodeCache = require( "node-cache" )
+const cache = new NodeCache( { stdTTL: 100, checkperiod: 120 } )
+
+module.exports = {
+    cache
+}
