+++ slackapi-app/commands/index.js
@@ -1,0 +1,14 @@
+const { showUserInfo } = require('./showUserInfo')
+const { showUsersLog } = require('./showUsersLog')
+const { conbot } = require('./conbot')
+
+
+const initCommands = ({users, channels}) => {
+    showUserInfo(users, channels)
+    showUsersLog(users, channels)
+    conbot()
+}
+
+module.exports = {
+    initCommands
+}
