+++ slackapi-app/package.json
@@ -2,9 +2,9 @@
   "name": "bolt-js-getting-started-app",
   "version": "2.0.0",
   "description": "Getting Started ⚡️ Bolt for JavaScript",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "nodemon index.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "prisma-push": "prisma db push"
   },
